refactor: reorganize type system into dedicated files
- Move types from index.ts to dedicated type files for better organization
- Create common.types.ts for shared interfaces like ModuleInfo and ChangeInfo
- Create config.types.ts for configuration-related types
- Create context.types.ts for workspace context types
- Create github.types.ts for GitHub API related types
- Create wiki.types.ts for wiki-specific types
- Create node-child-process.types.ts for Node.js process types
- Update index.ts to re-export all types from dedicated files

This refactoring improves code organization and makes types easier to maintain.
